Confirmed balances have to be positive
h/t @sgeisler https://github.com/bitcoindevkit/rust-electrum-client/issues/45#issuecomment-779453735

+++ src/electrum/server.rs
@@ -286,7 +286,7 @@ impl Connection {
         let (chain_stats, mempool_stats) = self.query.stats(&script_hash[..]);
 
         Ok(json!({
-            "confirmed": chain_stats.funded_txo_sum as i64 - chain_stats.spent_txo_sum as i64,
+            "confirmed": chain_stats.funded_txo_sum - chain_stats.spent_txo_sum,
             "unconfirmed": mempool_stats.funded_txo_sum as i64 - mempool_stats.spent_txo_sum as i64,
         }))
     }
